add deploy script
