--- config/scripts/build.ts
@@ -1,20 +1,0 @@
-function check() {
-
-}
-
-
-function main() {
-//  checkIfRelease();
-}
-
-main();
-
-/*
-
-build
-publish
-
-pre_integration_check()
-post_integration_check()
-
-*/
