+++ scripts/deploy.ts
@@ -1,16 +1,17 @@
-import * as child_process from "child_process";
 import * as fs from "fs";
 import * as stream from "stream";
-import {PACKAGES, getBuildInformation, print, print_error, run} from "./common";
+import {PACKAGES, print, print_error, run} from "./common";
+import {getBuildInformation} from "./release";
+import * as child_process from 'child_process';
+
+const conventionalChangelog = require("conventional-changelog");
 
 const RELEASE_TIMEOUT = 10 * 60 * 1000; // 10 minutes
+let RELEASE_BRANCH = "";
 
 const GH_TOKEN = process.env.GH_TOKEN;
 const COMMIT_TAG = process.env.TRAVIS_TAG;
 const TRAVIS_REPO_SLUG = process.env.TRAVIS_REPO_SLUG;
-
-const BUILD_INFO = getBuildInformation();
-const RELEASE_BRANCH = "Releasing_" + BUILD_INFO.release;
 
 const CHANGELOG = {
   file: "CHANGELOG.md",
@@ -37,20 +38,21 @@ const CHANGELOG = {
   }
 };
 
+const BUILD_INFO = getBuildInformation();
+
 main();
 
 function main() {
   if (BUILD_INFO.release) {
+    RELEASE_BRANCH = "Releasing_" + BUILD_INFO.release;
+
     updateChangelog()
       .then(() => prepareReleaseBranch())
       .then(() => delayRelease())
+      .then(() => loginToNPM())
       .then(() => {
         publishToNPM();
         mergeReleaseBranch();
-      })
-      .catch((e) => {
-        print_error(e);
-        process.exit(1);
       });
   } else {
     print("Nothing to deploy...");
@@ -59,6 +61,7 @@ function main() {
 
 function loginToNPM() {
   return new Promise((resolve, reject) => {
+    // Login to npm.
     const username = process.env.NPM_USERNAME;
     const password = process.env.NPM_PASSWORD;
     const email = process.env.NPM_EMAIL;
@@ -97,32 +100,26 @@ function loginToNPM() {
 function publishToNPM() {
   print(`====== PUBLISHING: Version ${BUILD_INFO.version}`);
 
-  loginToNPM()
-    .then(() => {
-      for (const PACKAGE of PACKAGES) {
-        run("npm", ["publish", `${process.cwd()}/dist/packages-dist/${PACKAGE}`, "--access=public"]);
-        print(`======      [${PACKAGE}]: PUBLISHED =====`);
-      }
-    });
+  for (const PACKAGE of PACKAGES) {
+    // run("npm", ["publish", `${ROOT_DIR}/dist/packages-dist/${PACKAGE}`, "--access=public"]);
+    print(`======      [${PACKAGE}]: PUBLISHED =====`);
+  }
 }
 
-function updateChangelog() {
-  print("====== Update changelog.");
-  return generateChangelog()
-    .then(() => {
-      if (fs.existsSync(CHANGELOG.file)) {
-        const data = fs.readFileSync(CHANGELOG.file);
-        fs.writeFileSync(CHANGELOG.file, CHANGELOG.log + data);
-      } else {
-        fs.writeFileSync(CHANGELOG.file, CHANGELOG.log);
-      }
-    });
+async function updateChangelog() {
+  await generateChangelog();
+  if (fs.existsSync(CHANGELOG.file)) {
+    const data = fs.readFileSync(CHANGELOG.file);
+    fs.writeFileSync(CHANGELOG.file, CHANGELOG.log + data);
+  } else {
+    fs.writeFileSync(CHANGELOG.file, CHANGELOG.log);
+  }
 }
 
 function generateChangelog() {
   return new Promise((resolve, reject) => {
     // Initiate the source
-    const changelog_buffer = new stream.PassThrough();
+    let changelog_buffer = new stream.PassThrough();
 
     let changelog = "";
     changelog_buffer.on("data", (data) => {
@@ -134,13 +131,13 @@ function generateChangelog() {
     });
 
     changelog_buffer.on("end", () => {
-      const changelog_lines = changelog.split("\n");
+      let changelog_lines = changelog.split("\n");
 
       CHANGELOG.header = changelog_lines.splice(0, 2).join("\n");
       CHANGELOG.changes = changelog_lines.splice(2, changelog_lines.length - 6).join("\n");
 
       // Get changelog body.
-      const tag_message = run("git", ["cat-file", "-p", COMMIT_TAG])[1].toString().split("\n");
+      let tag_message = run("git", ["cat-file", "-p", COMMIT_TAG])[1].toString().split("\n");
       if (tag_message[0].startsWith("object")) {
         CHANGELOG.body = tag_message.slice(5).join("\n");
       }
@@ -149,7 +146,7 @@ function generateChangelog() {
     });
 
     // Generate changelog from commit messages.
-    require("conventional-changelog")({
+    conventionalChangelog({
       preset: "angular"
     }).pipe(changelog_buffer);
   });
@@ -175,7 +172,7 @@ function delayRelease() {
 }
 
 function prepareReleaseBranch() {
-  print("====== Prepare release branch");
+  if (!process.env.ABC) return;
   run("git", ["checkout", "-b", RELEASE_BRANCH]);
   run("git", ["add", "-u"]);
   run("git", ["add", "dist/packages/*"]);
@@ -185,7 +182,7 @@ function prepareReleaseBranch() {
 }
 
 function mergeReleaseBranch() {
-  print("====== Merge release branch");
+  if (!process.env.ABC) return;
   run("git", ["checkout", "master"]);
   run("git", ["merge", RELEASE_BRANCH]);
 
@@ -200,7 +197,8 @@ function mergeReleaseBranch() {
   run("git", ["branch", "-d", RELEASE_BRANCH]);
   run("git", ["push", "origin", "--delete", RELEASE_BRANCH]);
 
-  print("====== Create GitHub release");
+  print("====== Create GitHub release documentation");
+
   const release = {
     "tag_name": BUILD_INFO.version,
     "target_commitish": "master",
